Split player limit service into vault and non vault implementations
When vault isn't installed, the plugin should fall back to an
alternative player limit service that doesn't rely on specialised vault
connections.
